+++ src/main/java/me/foxikle/foxrank/FoxRank.java
@@ -7,12 +7,7 @@ import org.bukkit.ChatColor;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.player.AsyncPlayerChatEvent;
-import org.bukkit.event.player.AsyncPlayerPreLoginEvent;
-import org.bukkit.event.player.PlayerJoinEvent;
-import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scoreboard.Scoreboard;
 import org.bukkit.scoreboard.Team;
@@ -116,6 +111,7 @@ public class FoxRank extends JavaPlugin implements Listener {
         getServer().getPluginManager().registerEvents(new Vanish(), this);
         getServer().getPluginManager().registerEvents(new JoinLeaveMsgs(), this);
         getServer().getPluginManager().registerEvents(new Logs(), this);
+        getServer().getPluginManager().registerEvents(new Listeners(), this);
         reloadConfig();
         for (Player p : this.getServer().getOnlinePlayers()) {
             db.addPlayerData(new RankedPlayer(p));
@@ -153,7 +149,7 @@ public class FoxRank extends JavaPlugin implements Listener {
         db.disconnect();
     }
 
-    private void setupTeams() {
+    void setupTeams() {
         Scoreboard board = Bukkit.getScoreboardManager().getMainScoreboard();
         if (!disableRankVis) {
             try {
@@ -398,119 +394,6 @@ public class FoxRank extends JavaPlugin implements Listener {
         ranks.remove(player);
     }
 
-    @EventHandler
-    public void onLeave(PlayerQuitEvent event) {
-        Player p = event.getPlayer();
-        saveRank(p);
-    }
-
-    @EventHandler
-    public void onPlayerChat(AsyncPlayerChatEvent e) {
-        e.setCancelled(true);
-        String eventMessage = e.getMessage();
-        String newMessage;
-        Player player = e.getPlayer();
-        UUID uuid = player.getUniqueId();
-
-        if (useDb) {
-            if (isMuted(uuid)) {
-                Instant date = Instant.parse(db.getStoredMuteDuration(uuid));
-                Instant now = Instant.now();
-                if (date.isBefore(now)) {
-                    unmutePlayer(new RankedPlayer(player), new RankedPlayer(player));
-                } else {
-                    String reason = db.getStoredMuteReason(uuid);
-                    String border = RED + "" + STRIKETHROUGH + "                                                                   ";
-                    String muteMessage = FoxRank.getInstance().getConfig().getString("ChatWhileMutedMessage").replace("$LINE", border);
-                    muteMessage = muteMessage.replace("\\n", "\n");
-                    muteMessage = muteMessage.replace("$DURATION", getFormattedExpiredString(date, Instant.now()));
-                    muteMessage = muteMessage.replace("$REASON", reason);
-                    muteMessage = ChatColor.translateAlternateColorCodes('§', muteMessage);
-                    player.sendMessage(muteMessage);
-                    return;
-                }
-            }
-            if (db.getStoredNicknameStatus(uuid)) {
-                Rank rank = db.getStoredNicknameRank(uuid);
-                String nick = db.getStoredNickname(uuid);
-                if (!disableRankVis) {
-                    if (rank == DEFAULT) {
-                        newMessage = ChatColor.GRAY + nick + ": " + eventMessage;
-                        Bukkit.broadcastMessage(newMessage);
-                    } else {
-                        newMessage = rank.getPrefix() + "" + rank + ChatColor.RESET + ": " + eventMessage;
-                        Bukkit.broadcastMessage(newMessage);
-                    }
-                } else {
-                    Bukkit.broadcastMessage(nick + ": " + e.getMessage());
-                }
-            } else {
-
-                if (!disableRankVis) {
-                    if (getRank(player) != DEFAULT) {
-                        newMessage = getRank(player).getPrefix() + player.getName() + ChatColor.RESET + ": " + eventMessage;
-                        Bukkit.broadcastMessage(newMessage);
-                    } else if (getRank(player) == DEFAULT) {
-                        newMessage = getRank(player).getPrefix() + player.getName() + ChatColor.RESET + "" + ChatColor.GRAY + ": " + eventMessage;
-                        Bukkit.broadcastMessage(newMessage);
-                    }
-                } else {
-                    Bukkit.broadcastMessage(player.getName() + ": " + e.getMessage());
-                }
-            }
-        } else {
-            File file = new File("plugins/FoxRank/PlayerData/" + player.getUniqueId() + ".yml");
-            YamlConfiguration yml = YamlConfiguration.loadConfiguration(file);
-
-            if (isMuted(player.getUniqueId())) {
-                Instant date = Instant.parse(yml.getString("MuteDuration"));
-                Instant now = Instant.now();
-                if (date.isBefore(now)) {
-                    unmutePlayer(new RankedPlayer(player), new RankedPlayer(player));
-                } else {
-                    String reason = yml.getString("MuteReason");
-                    String border = RED + "" + STRIKETHROUGH + "                                                                   ";
-                    String muteMessage = FoxRank.getInstance().getConfig().getString("ChatWhileMutedMessage").replace("$LINE", border);
-                    muteMessage = muteMessage.replace("\\n", "\n");
-                    muteMessage = muteMessage.replace("$DURATION", getFormattedExpiredString(date, Instant.now()));
-                    muteMessage = muteMessage.replace("$REASON", reason);
-                    muteMessage = ChatColor.translateAlternateColorCodes('§', muteMessage);
-                    player.sendMessage(muteMessage);
-                    return;
-                }
-            }
-            if (yml.getBoolean("isNicked")) {
-                Rank rank = ofString(yml.getString("Nickname-Rank"));
-                String nick = yml.getString("Nickname");
-                if (!disableRankVis) {
-
-                    if (rank == DEFAULT) {
-                        newMessage = ChatColor.GRAY + nick + ": " + eventMessage;
-                        Bukkit.broadcastMessage(newMessage);
-                    } else {
-                        newMessage = rank.getPrefix() + "" + nick + ChatColor.RESET + ": " + eventMessage;
-                        Bukkit.broadcastMessage(newMessage);
-                    }
-                } else {
-                    Bukkit.broadcastMessage(nick + ": " + e.getMessage());
-                }
-            } else {
-
-                if (!disableRankVis) {
-                    if (getRank(player) != DEFAULT) {
-                        newMessage = getRank(player).getPrefix() + player.getName() + ChatColor.RESET + ": " + eventMessage;
-                        Bukkit.broadcastMessage(newMessage);
-                    } else if (getRank(player) == DEFAULT) {
-                        newMessage = getRank(player).getPrefix() + player.getName() + ChatColor.RESET + "" + ChatColor.GRAY + ": " + eventMessage;
-                        Bukkit.broadcastMessage(newMessage);
-                    }
-                } else {
-                    Bukkit.broadcastMessage(player.getName() + ": " + e.getMessage());
-                }
-            }
-        }
-    }
-
     protected String getFormattedExpiredString(Instant date, Instant now) {
         if (date == null) {
             return "Permanant";
@@ -537,54 +420,6 @@ public class FoxRank extends JavaPlugin implements Listener {
                 str = str + seconds + "s";
             }
             return str;
-        }
-    }
-
-    @EventHandler
-    public void OnPlayerLogin(PlayerJoinEvent e) {
-        setupTeams();
-        Player p = e.getPlayer();
-        if (useDb) {
-            db.addPlayerData(new RankedPlayer(p));
-        } else {
-
-            File file = new File("plugins/FoxRank/PlayerData/" + p.getUniqueId() + ".yml");
-            if (!file.exists()) {
-                try {
-                    file.createNewFile();
-                } catch (IOException error) {
-                    error.printStackTrace();
-                }
-            }
-            YamlConfiguration yml = YamlConfiguration.loadConfiguration(file);
-            yml.addDefault("Name", p.getName());
-            yml.addDefault("UUID", p.getUniqueId().toString());
-            yml.addDefault("Rank", "DEFAULT");
-            yml.addDefault("isVanished", false);
-            yml.addDefault("isNicked", false);
-            yml.addDefault("isMuted", false);
-            yml.addDefault("MuteDuration", "");
-            yml.addDefault("MuteReason", "");
-            yml.addDefault("Nickname", p.getName());
-            yml.addDefault("Nickname-Rank", "DEFAULT");
-            yml.addDefault("Nickname-Skin", "");
-            yml.addDefault("BanDuration", "");
-            yml.addDefault("BanReason", "");
-            yml.addDefault("BanID", "");
-            yml.addDefault("isBanned", false);
-            yml.options().copyDefaults(true);
-            try {
-                yml.save(file);
-            } catch (IOException error) {
-                error.printStackTrace();
-            }
-        }
-        ActionBar.setupActionBar(p);
-        loadRank(p);
-        if (isMuted(p.getUniqueId())) {
-            if (getMuteDuration(p.getUniqueId()).isBefore(Instant.now())) {
-                unmutePlayer(new RankedPlayer(p), new RankedPlayer(p));
-            }
         }
     }
 
@@ -848,73 +683,6 @@ public class FoxRank extends JavaPlugin implements Listener {
         return uid;
     }
 
-    @EventHandler
-    public void BanHandler(AsyncPlayerPreLoginEvent e) {
-        if (isBanned(e.getUniqueId())) {
-            UUID uuid = e.getUniqueId();
-            String bumper = "\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n";
-            String reason;
-            String duration;
-            String id;
-            if (useDb) {
-                reason = db.getStoredBanReason(uuid);
-                duration = db.getStoredBanDuration(uuid);
-                id = db.getStoredBanID(uuid);
-            } else {
-
-                File file = new File("plugins/FoxRank/PlayerData/" + uuid + ".yml");
-                YamlConfiguration yml = YamlConfiguration.loadConfiguration(file);
-                reason = yml.getString("BanReason");
-                duration = yml.getString("BanDuration");
-                id = yml.getString("BanID");
-            }
-
-            if (duration == null) {
-                String finalMessage = this.getConfig().getString("PermBanMessageFormat")
-                        .replace("$SERVER_NAME", this.getConfig().getString("ServerName"))
-                        .replace("$REASON", reason)
-                        .replace("$APPEAL_LINK", this.getConfig().getString("BanAppealLink"))
-                        .replace("$ID", id)
-                        .replace("\\n", "\n");
-                e.disallow(AsyncPlayerPreLoginEvent.Result.KICK_BANNED, ChatColor.translateAlternateColorCodes('§', bumper + finalMessage + bumper));
-            } else {
-                Instant inst = Instant.parse(duration);
-                if (Instant.now().isAfter(inst)) {
-                    unbanOfflinePlayer(uuid, null);
-
-                    if (useDb) {
-
-                    } else {
-                        File bannedPlayersFile = new File("plugins/FoxRank/bannedPlayers.yml");
-                        YamlConfiguration bannedPlayersyml = YamlConfiguration.loadConfiguration(bannedPlayersFile);
-                        List<String> list = bannedPlayersyml.getStringList("CurrentlyBannedPlayers");
-                        if (list.contains(e.getUniqueId().toString())) {
-                            if (list.remove(e.getUniqueId().toString())) {
-                                bannedPlayersyml.set("CurrentlyBannedPlayers", list);
-                                try {
-                                    bannedPlayersyml.save(bannedPlayersFile);
-                                } catch (IOException ex) {
-                                    ex.printStackTrace();
-                                }
-                            }
-                        }
-                    }
-                    e.allow();
-                    return;
-                }
-                String finalMessage = this.getConfig().getString("TempBanMessageFormat")
-                        .replace("$DURATION", getFormattedExpiredString(inst, Instant.now()))
-                        .replace("$SERVER_NAME", this.getConfig().getString("ServerName"))
-                        .replace("$REASON", reason)
-                        .replace("$APPEAL_LINK", this.getConfig().getString("BanAppealLink"))
-                        .replace("$ID", id)
-                        .replace("\\n", "\n");
-                e.disallow(AsyncPlayerPreLoginEvent.Result.KICK_BANNED, ChatColor.translateAlternateColorCodes('§', bumper + finalMessage + bumper));
-            }
-        } else {
-            e.allow();
-        }
-    }
 
     public List<OfflinePlayer> getBannedPlayers() {
         if (useDb) {
